Fix tests, cleanup dashboard resource
