+++ ui/src/app/data-explorer/components/chart-view/designer-panel/data-settings/field-selection-panel/field-selection-panel.component.html
@@ -55,16 +55,16 @@
             <mat-icon>expand_less</mat-icon>
         </button>
     </div>
-    <div fxLayout="column" [class.field-scroll-panel]="!expandFields">
+    <div
+        fxLayout="column"
+        [class.field-scroll-panel]="!expandFields"
+        data-cy="data-explorer-field-selection-panel"
+    >
         <div
             fxLayout="column"
             *ngFor="let field of sourceConfig.queryConfig.fields"
         >
-            <sp-field-selection
-                [field]="field"
-                [sourceConfig]="sourceConfig"
-                [widgetId]="widgetId"
-            >
+            <sp-field-selection [field]="field" [sourceConfig]="sourceConfig">
             </sp-field-selection>
         </div>
     </div>
